requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
openpyxl==3.1.3
pandas==2.2.2
cryptography==42.0.8
python-dotenv==1.0.1
Werkzeug==3.0.3
